server/requirements.txt
Flask
langchain
langchain-community
openai
torch
torchvision
transformers
pillow